refactor(Select): update remote search example to use TypeScript types
